pdApp/requirements.txt
Flask==2.2.2
tensorflow==2.11.0
opencv-python==4.7.0.72
numpy==1.24.2
pickleDB==0.9.2
Pillow==9.4.0
requests==2.28.2